% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d3rug.R
\name{d3rug}
\alias{d3rug}
\title{Interactive Rug plot}
\usage{
d3rug(values, labels = names(values), col = "steelblue", col_hover = col,
  alpha = 0.3, unit = NULL, group = NULL, select = NULL,
  col_select = "red", main = deparse(substitute(values)), ...,
  width = "100\%", height = "150px")
}
\arguments{
\item{values}{\code{numeric} with values to be plotted}

\item{labels}{\code{character} of same lenght as \code{values}. These will be used to
annotate a value when hovered.}

\item{col}{\code{character} with colors. Either a scalar or length of \code{values}.}

\item{col_hover}{\code{character} with colors used for coloring hovered values.}

\item{alpha}{transparancy of values which are not hovered or selected.}

\item{unit}{\code{character} text to be appended to value.}

\item{group}{optional \code{character} that is used to highlight the group of
values that are hovered or selected}

\item{select}{optional \code{character} specifying which label is selected}

\item{col_select}{\code{character} with color used for selected values.}

\item{main}{title of the plot}

\item{...}{extra parameters}

\item{width}{width of plot}

\item{height}{height of plot}
}
\description{
\code{d3rug} creates an interactive rug plot.
The rugplot is heavily inspired by the Lambrechts Strip Plot (reference).
}
\examples{

\dontrun{
library(d3rug)
income <- state.x77[,2]
d3rug(income, height="100px")
illiteracy_levels <- cut(state.x77[,3], 5)
pal <- colorRampPalette(c("green","gray", "orange"))(nlevels(illiteracy_levels))
col = pal[illiteracy_levels]
d3rug(income, color=col)
}
}

