% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d3rug.R
\name{d3rug}
\alias{d3rug}
\title{Interactive Rug plot}
\usage{
d3rug(message, width = NULL, height = NULL)
}
\description{
\code{d3rug} creates an interactive rug plot.
The rugplot is heavily inspired by the Lambrechts Strip Plot (reference).
}

